TST: fix setup and teardown in db-related tests
do not execute them if the INTELMQ_TEST_DATABASES is not set

+++ intelmq/tests/bots/outputs/postgresql/test_output.py
@@ -53,6 +53,8 @@ class TestPostgreSQLOutputBot(test.BotTestCase, unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
+        if not os.environ.get('INTELMQ_TEST_DATABASES'):
+            return
         cls.cur.execute('TRUNCATE "tests"')
         cls.cur.close()
         cls.con.close()
